--- a/e2e_test.go
+++ b/e2e_test.go
@@ -18,7 +18,6 @@ import (
 )
 
 func TestE2E(t *testing.T) {
-	t.Parallel()
 	suite.Run(t, &e2eSuite{})
 }