--- a/router.go
+++ b/router.go
@@ -8,57 +8,49 @@ import (
 	"github.com/monzo/terrors"
 )
 
-type Router interface {
-	// OPTIONS is a shortcut for Register("OPTIONS", path, svc).
-	OPTIONS(pattern string, svc Service)
-	// GET is a shortcut for Register("GET", path, svc).
-	GET(pattern string, svc Service)
-	// HEAD is a shortcut for Register("HEAD", path, svc).
-	HEAD(pattern string, svc Service)
-	// POST is a shortcut for Register("POST", path, svc).
-	POST(pattern string, svc Service)
-	// PUT is a shortcut for Register("PUT", path, svc).
-	PUT(pattern string, svc Service)
-	// DELETE is a shortcut for Register("DELETE", path, svc).
-	DELETE(pattern string, svc Service)
-	// TRACE is a shortcut for Register("TRACE", path, svc).
-	TRACE(pattern string, svc Service)
-	// Register associates a Service with a method and path.
-	Register(method, pattern string, svc Service)
-	// Lookup returns the Service, pattern, and extracted path parameters for the HTTP method and path.
-	Lookup(method, path string) (svc Service, pattern string, params map[string]string, ok bool)
-	// Serve returns a Service which will route inbound requests to the enclosed routes.
-	Serve() Service
-	// Pattern returns the registered pattern which matches the given request.
-	Pattern(req Request) string
-	// Params returns extracted URL parameters, assuming the request has been routed and has captured parameters.
-	Params(req Request) map[string]string
-}
-
-type router struct {
+// A Router multiplexes requests to a set of Services by pattern matching on method and path, and can also extract
+// parameters from paths.
+type Router struct {
 	e    *echo.Echo
 	r    *echo.Router
 	svcs map[string]Service
-	m    sync.RWMutex
+	m    *sync.RWMutex
 }
 
 // NewRouter vends a new implementation of Router
 func NewRouter() Router {
 	e := echo.New()
-	return &router{
+	return Router{
 		e:    e,
 		r:    echo.NewRouter(e),
-		svcs: make(map[string]Service, 10)}
+		svcs: make(map[string]Service, 10),
+		m:    new(sync.RWMutex)}
 }
 
-func (r *router) Register(method, pattern string, svc Service) {
+// Register associates a Service with a method and path.
+//
+// Method is a single HTTP method name, or * which is expanded to {OPTIONS, GET, HEAD, POST, PUT, DELETE, TRACE}.
+// Pattern syntax is as described in echo's documentation: https://echo.labstack.com/guide/routing
+func (r *Router) Register(method, pattern string, svc Service) {
+	echoHandler := func(c echo.Context) error { return nil }
+
 	r.m.Lock()
 	defer r.m.Unlock()
-	r.r.Add(method, pattern, func(c echo.Context) error { return nil })
-	r.svcs[method+pattern] = svc
+
+	if method == "*" {
+		// Expand * to the set of all known methods
+		for _, m := range [...]string{"GET", "CONNECT", "DELETE", "HEAD", "OPTIONS", "PATCH", "POST", "PUT", "TRACE"} {
+			r.r.Add(m, pattern, echoHandler)
+			r.svcs[m+pattern] = svc
+		}
+	} else {
+		r.r.Add(method, pattern, echoHandler)
+		r.svcs[method+pattern] = svc
+	}
 }
 
-func (r *router) Lookup(method, path string) (Service, string, map[string]string, bool) {
+// Lookup returns the Service, pattern, and extracted path parameters for the HTTP method and path.
+func (r *Router) Lookup(method, path string) (Service, string, map[string]string, bool) {
 	c := r.e.AcquireContext()
 	defer r.e.ReleaseContext(c)
 	c.Reset(nil, nil)
@@ -86,7 +78,8 @@ func (r *router) Lookup(method, path string) (Service, string, map[string]string
 	return svc, pattern, params, true
 }
 
-func (r *router) Serve() Service {
+// Serve returns a Service which will route inbound requests to the enclosed routes.
+func (r *Router) Serve() Service {
 	return func(req Request) Response {
 		svc, _, _, ok := r.Lookup(req.Method, req.URL.Path)
 		if !ok {
@@ -99,21 +92,61 @@ func (r *router) Serve() Service {
 	}
 }
 
-func (r *router) Pattern(req Request) string {
+// Pattern returns the registered pattern which matches the given request.
+func (r *Router) Pattern(req Request) string {
 	_, pattern, _, _ := r.Lookup(req.Method, req.URL.Path)
 	return pattern
 }
 
-func (r *router) Params(req Request) map[string]string {
+// Params returns extracted path parameters, assuming the request has been routed and has captured parameters.
+func (r *Router) Params(req Request) map[string]string {
 	_, _, params, _ := r.Lookup(req.Method, req.URL.Path)
 	return params
 }
 
 // Sugar
-func (r *router) OPTIONS(pattern string, svc Service) { r.Register("OPTIONS", pattern, svc) }
-func (r *router) GET(pattern string, svc Service)     { r.Register("GET", pattern, svc) }
-func (r *router) HEAD(pattern string, svc Service)    { r.Register("HEAD", pattern, svc) }
-func (r *router) POST(pattern string, svc Service)    { r.Register("POST", pattern, svc) }
-func (r *router) PUT(pattern string, svc Service)     { r.Register("PUT", pattern, svc) }
-func (r *router) DELETE(pattern string, svc Service)  { r.Register("DELETE", pattern, svc) }
-func (r *router) TRACE(pattern string, svc Service)   { r.Register("TRACE", pattern, svc) }
+
+// GET is shorthand for Register("GET", pattern, svc).
+//
+// Pattern syntax is as described in echo's documentation: https://echo.labstack.com/guide/routing
+func (r *Router) GET(pattern string, svc Service) { r.Register("GET", pattern, svc) }
+
+// CONNECT is shorthand for Register("CONNECT", pattern, svc).
+//
+// Pattern syntax is as described in echo's documentation: https://echo.labstack.com/guide/routing
+func (r *Router) CONNECT(pattern string, svc Service) { r.Register("CONNECT", pattern, svc) }
+
+// DELETE is shorthand for Register("DELETE", pattern, svc).
+//
+// Pattern syntax is as described in echo's documentation: https://echo.labstack.com/guide/routing
+func (r *Router) DELETE(pattern string, svc Service) { r.Register("DELETE", pattern, svc) }
+
+// HEAD is shorthand for Register("HEAD", pattern, svc).
+//
+// Pattern syntax is as described in echo's documentation: https://echo.labstack.com/guide/routing
+func (r *Router) HEAD(pattern string, svc Service) { r.Register("HEAD", pattern, svc) }
+
+// OPTIONS is shorthand for Register("OPTIONS", pattern, svc).
+//
+// Pattern syntax is as described in echo's documentation: https://echo.labstack.com/guide/routing
+func (r *Router) OPTIONS(pattern string, svc Service) { r.Register("OPTIONS", pattern, svc) }
+
+// PATCH is shorthand for Register("PATCH", pattern, svc).
+//
+// Pattern syntax is as described in echo's documentation: https://echo.labstack.com/guide/routing
+func (r *Router) PATCH(pattern string, svc Service) { r.Register("PATCH", pattern, svc) }
+
+// POST is shorthand for Register("POST", pattern, svc).
+//
+// Pattern syntax is as described in echo's documentation: https://echo.labstack.com/guide/routing
+func (r *Router) POST(pattern string, svc Service) { r.Register("POST", pattern, svc) }
+
+// PUT is shorthand for Register("PUT", pattern, svc).
+//
+// Pattern syntax is as described in echo's documentation: https://echo.labstack.com/guide/routing
+func (r *Router) PUT(pattern string, svc Service) { r.Register("PUT", pattern, svc) }
+
+// TRACE is shorthand for Register("TRACE", pattern, svc).
+//
+// Pattern syntax is as described in echo's documentation: https://echo.labstack.com/guide/routing
+func (r *Router) TRACE(pattern string, svc Service) { r.Register("TRACE", pattern, svc) }