`Router` is a fairly crucial piece of functionality in Typhon, used internally at Monzo in basically every backend service – often multiple times. However it's also poorly tested and has subtle behaviour that can be hard to understand.

## `Router` is no longer an interface

This change simplifies `Router` somewhat so it's no longer an interface. The rest of Typhon doesn't really care about `Router`s – it only sees the `Service`s that are materialised from them – so it didn't make sense to have an interface adding a layer of indirection. Calling code should not need to be modified so this is not a breaking change.

## Catch-all method sugar

`"*"` can now be passed as a method name to `Register` if a caller wants all methods for a given path to be routed to a single service. Under the hood, `*` is expanded to a set of {`OPTIONS`, `GET`, `HEAD`, `POST`, `PUT`, `DELETE`, `TRACE`}.

## Housekeeping

`Router`'s methods are better-documented, there are more test cases, and now benchmarks too.

## Performance

A number of small improvements mean that routing requests now often involves zero memory allocations, and is substantially faster:

```
benchmark                                                  old allocs     new allocs     delta
BenchmarkRouter/Serve/GET/-4                               20             20             +0.00%
BenchmarkRouter/Serve/GET/foo-4                            1              0              -100.00%
BenchmarkRouter/Serve/GET/foo/bar2bär/baz-4                2              0              -100.00%
BenchmarkRouter/Serve/GET/foo/bar/bar/baz-4                20             20             +0.00%
BenchmarkRouter/Serve/GET/residual/r-4                     2              0              -100.00%
BenchmarkRouter/Serve/GET/residual/r/e/s/i/d/u/a/l-4       2              0              -100.00%
BenchmarkRouter/Serve/GET/residual/r/e/s/i/d/u/a/l/-4      2              0              -100.00%
BenchmarkRouter/Serve/WTAF/poly-4                          20             20             +0.00%
BenchmarkRouter/Serve/GET/poly-4                           20             0              -100.00%
BenchmarkRouter/Serve/CONNECT/poly-4                       20             0              -100.00%
BenchmarkRouter/Serve/DELETE/poly-4                        20             0              -100.00%
BenchmarkRouter/Serve/HEAD/poly-4                          20             0              -100.00%
BenchmarkRouter/Serve/OPTIONS/poly-4                       20             0              -100.00%
BenchmarkRouter/Serve/PATCH/poly-4                         20             0              -100.00%
BenchmarkRouter/Serve/POST/poly-4                          20             0              -100.00%
BenchmarkRouter/Serve/PUT/poly-4                           20             0              -100.00%
BenchmarkRouter/Serve/TRACE/poly-4                         20             0              -100.00%
```

```
benchmark                                                  old ns/op     new ns/op     delta
BenchmarkRouter/Serve/GET/-4                               4277          4156          -2.83%
BenchmarkRouter/Serve/GET/foo-4                            302           216           -28.48%
BenchmarkRouter/Serve/GET/foo/bar2bär/baz-4                785           273           -65.22%
BenchmarkRouter/Serve/GET/foo/bar/bar/baz-4                14729         4602          -68.76%
BenchmarkRouter/Serve/GET/residual/r-4                     684           243           -64.47%
BenchmarkRouter/Serve/GET/residual/r/e/s/i/d/u/a/l-4       513           250           -51.27%
BenchmarkRouter/Serve/GET/residual/r/e/s/i/d/u/a/l/-4      565           241           -57.35%
BenchmarkRouter/Serve/WTAF/poly-4                          4821          4869          +1.00%
BenchmarkRouter/Serve/GET/poly-4                           4215          225           -94.66%
BenchmarkRouter/Serve/CONNECT/poly-4                       5496          193           -96.49%
BenchmarkRouter/Serve/DELETE/poly-4                        4363          223           -94.89%
BenchmarkRouter/Serve/HEAD/poly-4                          4525          308           -93.19%
BenchmarkRouter/Serve/OPTIONS/poly-4                       5198          209           -95.98%
BenchmarkRouter/Serve/PATCH/poly-4                         5039          230           -95.44%
BenchmarkRouter/Serve/POST/poly-4                          4490          266           -94.08%
BenchmarkRouter/Serve/PUT/poly-4                           4262          226           -94.70%
BenchmarkRouter/Serve/TRACE/poly-4                         4185          230           -94.50%
```